automatizacion-web/requirements.txt
selenium==4.24.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
requests==2.32.3
